requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
gradio>=4.0.0
python-dotenv>=1.0.0
requests>=2.25.0
wikipedia>=1.4.0
ollama>=0.1.0
pytest>=7.0.0